+++ src/common/strlib.cpp
@@ -1021,7 +1021,7 @@ bool sv_readdb(const char* directory, const char* filename, char delim, int minc
 		if( line[0] == '\0' || line[0] == '\n' || line[0] == '\r')
 			continue;
 
-		columns = sv_split(line, strlen(line), 0, delim, fields, nb_cols, (e_svopt)(SV_TERMINATE_LF|SV_TERMINATE_CRLF));
+		columns = sv_split(line, (int)strlen(line), 0, delim, fields, nb_cols, (e_svopt)(SV_TERMINATE_LF|SV_TERMINATE_CRLF));
 
 		if( columns < mincols )
 		{
@@ -1103,7 +1103,7 @@ int _StringBuf_Vprintf(const char *file, int line, const char *func,StringBuf* s
 		int n, size, off;
 		va_list apcopy;
 		/* Try to print in the allocated space. */
-		size = self->max_ - (self->ptr_ - self->buf_);
+		size = (int)(self->max_ - (self->ptr_ - self->buf_));
 		va_copy(apcopy, ap);
 		n = vsnprintf(self->ptr_, size, fmt, apcopy);
 		va_end(apcopy);
@@ -1124,7 +1124,7 @@ int _StringBuf_Vprintf(const char *file, int line, const char *func,StringBuf* s
 /// Appends the contents of another StringBuf to the StringBuf
 int _StringBuf_Append(const char *file, int line, const char *func,StringBuf* self, const StringBuf* sbuf)
 {
-	int available = self->max_ - (self->ptr_ - self->buf_);
+	int available = (int)(self->max_ - (self->ptr_ - self->buf_));
 	int needed = (int)(sbuf->ptr_ - sbuf->buf_);
 
 	if( needed >= available )
@@ -1143,7 +1143,7 @@ int _StringBuf_Append(const char *file, int line, const char *func,StringBuf* se
 // Appends str to the StringBuf
 int _StringBuf_AppendStr(const char *file, int line, const char *func,StringBuf* self, const char* str)
 {
-	int available = self->max_ - (self->ptr_ - self->buf_);
+	int available = (int)(self->max_ - (self->ptr_ - self->buf_));
 	int needed = (int)strlen(str);
 
 	if( needed >= available )
