+++ src/common/utilities.hpp
@@ -137,7 +137,7 @@ namespace rathena {
 		template <typename K, typename V> V& umap_random( std::unordered_map<K, V>& map ){
 			auto it = map.begin();
 
-			std::advance( it, rnd_value( 0, map.size() - 1 ) );
+			std::advance( it, rnd_value( 0, (int32)(map.size() - 1 )) );
 
 			return it->second;
 		}
@@ -150,7 +150,7 @@ namespace rathena {
 		template <typename K> K &vector_random(std::vector<K> &vec) {
 			auto it = vec.begin();
 
-			std::advance(it, rnd_value(0, vec.size() - 1));
+			std::advance(it, rnd_value(0, (int32)(vec.size() - 1)));
 
 			return *it;
 		}
