+++ src/map/duel.cpp
@@ -196,10 +196,10 @@ bool duel_invite(const size_t did, struct map_session_data* sd, struct map_sessi
 	char output[256];
 	// " -- Player %s invites %s to duel --"
 	sprintf(output, msg_txt(sd,373), sd->status.name, target_sd->status.name);
-	clif_disp_message(&sd->bl, output, strlen(output), DUEL_WOS);
+	clif_disp_message(&sd->bl, output, (int)strlen(output), DUEL_WOS);
 	// "Blue -- Player %s invites you to PVP duel (@accept/@reject) --"
 	sprintf(output, msg_txt(sd,374), sd->status.name);
-	clif_broadcast((struct block_list *)target_sd, output, strlen(output)+1, BC_BLUE, SELF);
+	clif_broadcast((struct block_list *)target_sd, output, (int)strlen(output)+1, BC_BLUE, SELF);
 	return true;
 }
 
@@ -239,7 +239,7 @@ bool duel_leave(const size_t did, struct map_session_data* sd)
 	char output[256];
 	// " <- Player %s has left duel --"
 	sprintf(output, msg_txt(sd,375), sd->status.name);
-	clif_disp_message(&sd->bl, output, strlen(output), DUEL_WOS);
+	clif_disp_message(&sd->bl, output, (int)strlen(output), DUEL_WOS);
 	clif_map_property(&sd->bl, MAPPROPERTY_NOTHING, SELF);
 	return true;
 }
@@ -265,7 +265,7 @@ bool duel_accept(const size_t did, struct map_session_data* sd)
 	char output[256];
 	// " -> Player %s has accepted duel --"
 	sprintf(output, msg_txt(sd,376), sd->status.name);
-	clif_disp_message(&sd->bl, output, strlen(output), DUEL_WOS);
+	clif_disp_message(&sd->bl, output, (int)strlen(output), DUEL_WOS);
 	clif_map_property(&sd->bl, MAPPROPERTY_FREEPVPZONE, SELF);
 	//clif_misceffect2(&sd->bl, 159);
 	return true;
@@ -290,7 +290,7 @@ bool duel_reject(const size_t did, struct map_session_data* sd)
 	char output[256];
 	// " -- Player %s has rejected duel --"
 	sprintf(output, msg_txt(sd,377), sd->status.name);
-	clif_disp_message(&sd->bl, output, strlen(output), DUEL_WOS);
+	clif_disp_message(&sd->bl, output, (int)strlen(output), DUEL_WOS);
 	return true;
 }
